Test AsyncMap in local or clustered mode
